requirements.txt
flask==2.3.3
pandas==2.0.3
openpyxl==3.1.2
qrcode==7.4.2
Pillow==10.0.0
flask-socketio==5.3.6
razorpay==1.4.1
python-socketio==5.9.0
gunicorn==21.2.0
eventlet==0.33.3
